+++ pkg/ddc/efc/transform_image_test.go
@@ -23,7 +23,7 @@ import (
 func TestParseMasterImage(t *testing.T) {
 	engine := &EFCEngine{}
 	image, tag, imagePullPolicy := engine.parseMasterImage("", "", "")
-	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/eac-master" ||
+	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/efc-master" ||
 		tag != "update" || imagePullPolicy != "IfNotPresent" {
 		t.Errorf("unexpected err")
 	}
@@ -32,7 +32,7 @@ func TestParseMasterImage(t *testing.T) {
 func TestParseFuseImage(t *testing.T) {
 	engine := &EFCEngine{}
 	image, tag, imagePullPolicy := engine.parseFuseImage("", "", "")
-	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/eac-fuse" ||
+	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/efc-fuse" ||
 		tag != "update" || imagePullPolicy != "IfNotPresent" {
 		t.Errorf("unexpected err")
 	}
@@ -41,7 +41,7 @@ func TestParseFuseImage(t *testing.T) {
 func TestParseWorkerImage(t *testing.T) {
 	engine := &EFCEngine{}
 	image, tag, imagePullPolicy := engine.parseWorkerImage("", "", "")
-	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/eac-worker" ||
+	if image != "registry.cn-zhangjiakou.aliyuncs.com/nascache/efc-worker" ||
 		tag != "update" || imagePullPolicy != "IfNotPresent" {
 		t.Errorf("unexpected err")
 	}